requirements.txt
python-telegram-bot
requests
feedparser
schedule
